infra/requirements.txt
reportlab>=4.0.0
google-cloud-bigquery>=3.0.0
google-cloud-discoveryengine>=0.11.0
google-cloud-storage>=2.10.0
